+++ CHANGELOG.md
@@ -6,7 +6,10 @@
 ## 0.3.2-snapshot
 - プロジェクト名を swith-tex-styles に変更（複数形）。
 - make install-bin
-  スクリプトを .clj をとってケバブスタイルの switch-tex-styles で ~/bin にインストール。
+  スクリプトファイル `switch_tex_styles.clj` を
+  拡張子 .clj をとったケバブスタイルの switch-tex-styles で ~/bin にインストール。
+- make install は ~/bin に、学術図書フォルダへは make install-gaku とする。
+
 
 ## 0.3.1 - 2023-10-01
 - プロジェクト名 switch-tex-style に変更。